Implemented caching of XSL markup generation.

+++ src/main/java/eu/europa/ted/eforms/viewer/NoticeViewer.java
@@ -10,7 +10,9 @@ import java.nio.charset.Charset;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.text.MessageFormat;
 import java.util.Optional;
+import java.util.function.Supplier;
 import javax.xml.XMLConstants;
 import javax.xml.parsers.DocumentBuilder;
 import javax.xml.parsers.ParserConfigurationException;
@@ -36,6 +38,7 @@ import org.xml.sax.SAXException;
 import eu.europa.ted.eforms.sdk.SdkConstants;
 import eu.europa.ted.eforms.sdk.resource.SdkDownloader;
 import eu.europa.ted.eforms.sdk.resource.SdkResourceLoader;
+import eu.europa.ted.eforms.viewer.helpers.CacheHelper;
 import eu.europa.ted.eforms.viewer.helpers.CustomUriResolver;
 import eu.europa.ted.eforms.viewer.helpers.SafeDocumentBuilder;
 import eu.europa.ted.efx.EfxTranslator;
@@ -44,9 +47,6 @@ import net.sf.saxon.trace.TimingTraceListener;
 
 public class NoticeViewer {
   private static final Logger logger = LoggerFactory.getLogger(NoticeViewer.class);
-
-  private static final Path OUTPUT_FOLDER_HTML = Path.of("target", "output-html");
-  private static final Path OUTPUT_FOLDER_XSL = Path.of("target", "output-xsl");
 
   private NoticeViewer() {}
 
@@ -58,6 +58,7 @@ public class NoticeViewer {
    *        type
    * @param profileXslt If set to true, XSLT profiling will be enabled
    * @param sdkRootPath Path of the root SDK folder
+   * @param forceBuild Forces the re-creation of XSL files
    * @return The path of the generated HTML file
    *
    * @throws IOException If an error occurs during input or output
@@ -66,7 +67,8 @@ public class NoticeViewer {
    * @throws InstantiationException
    */
   public static Path generateHtml(final String language, final Path noticeXmlPath,
-      final Optional<String> viewIdOpt, final boolean profileXslt, final Path sdkRootPath)
+      final Optional<String> viewIdOpt, final boolean profileXslt, final Path sdkRootPath,
+      boolean forceBuild)
       throws IOException, SAXException, ParserConfigurationException, InstantiationException {
     logger.debug("noticeXmlPath={}", noticeXmlPath);
     Validate.notNull(noticeXmlPath, "Invalid path to notice: " + noticeXmlPath);
@@ -93,7 +95,7 @@ public class NoticeViewer {
     final String eformsSdkVersion = eformsSdkVersionOpt.get();
     logger.debug("noticeSubType={}, viewId={}, eformsSdkVersion={}", noticeSubType, viewId,
         eformsSdkVersion);
-    final Path xslPath = buildXsl(viewId, eformsSdkVersion, sdkRootPath);
+    final Path xslPath = buildXsl(viewId, eformsSdkVersion, sdkRootPath, forceBuild);
     final Path htmlPath = applyXslTransform(language, eformsSdkVersion, noticeXmlPath, xslPath,
         viewId, profileXslt, sdkRootPath);
     // Ensure the HTML can be parsed.
@@ -214,7 +216,8 @@ public class NoticeViewer {
     factory.setURIResolver(new CustomUriResolver(sdkVersion, sdkRootPath));
 
     if (profileXslt) {
-      final Path xsltProfilePath = OUTPUT_FOLDER_HTML.resolve(viewId + "-xslt_profile.html");
+      final Path xsltProfilePath =
+          NoticeViewerConstants.OUTPUT_FOLDER_HTML.resolve(viewId + "-xslt_profile.html");
       logger.info("XSLT profiling is enabled. The result can be found at: {}", xsltProfilePath);
 
       factory.setAttribute(FeatureKeys.TRACE_LISTENER_CLASS, TimingTraceListener.class.getName());
@@ -252,9 +255,9 @@ public class NoticeViewer {
     final Source xmlInput = new StreamSource(noticeXmlPath.toFile());
 
     // HTML as output of the transformation.
-    Files.createDirectories(OUTPUT_FOLDER_HTML);
+    Files.createDirectories(NoticeViewerConstants.OUTPUT_FOLDER_HTML);
 
-    final Path htmlPath = OUTPUT_FOLDER_HTML.resolve(viewId + ".html");
+    final Path htmlPath = NoticeViewerConstants.OUTPUT_FOLDER_HTML.resolve(viewId + ".html");
     final StreamResult outputTarget = new StreamResult(htmlPath.toFile());
     try (InputStream inputStream = Files.newInputStream(xslPath)) {
       final Source xslSource = new StreamSource(inputStream);
@@ -287,6 +290,24 @@ public class NoticeViewer {
     }
   }
 
+  private static Supplier<String> templateTranslator(final Path sdkRootPath,
+      final String sdkVersion,
+      final InputStream viewInputStream, final String viewId) {
+    return () -> {
+      try {
+        return EfxTranslator.translateTemplate(new DependencyFactory(sdkRootPath),
+            sdkVersion,
+            viewInputStream);
+      } catch (InstantiationException | IOException e) {
+        throw new RuntimeException(
+            MessageFormat.format(
+                "Failed to build XSL for view ID [{0}] and SDK version[{1}]",
+                viewId, sdkVersion),
+            e);
+      }
+    };
+  }
+
   /**
    * Takes the EFX view template as a viewId string and outputs the XSL.
    *
@@ -294,11 +315,13 @@ public class NoticeViewer {
    *        from SDK by using naming conventions
    * @param sdkVersion The version of the desired SDK
    * @param sdkRootPath Path of the root SDK folder
+   * @param forceBuild Forces the re-creation of XSL files
    * @return Path to the built file
    * @throws IOException If an error occurred while writing the file
    * @throws InstantiationException
    */
-  public static final Path buildXsl(final String viewId, final String sdkVersion, Path sdkRootPath)
+  public static final Path buildXsl(final String viewId, final String sdkVersion, Path sdkRootPath,
+      boolean forceBuild)
       throws IOException, InstantiationException {
     logger.debug("Creating XSL for view ID [{}] and SDK version [{}]", viewId, sdkVersion);
 
@@ -306,24 +329,36 @@ public class NoticeViewer {
 
     logger.debug("View path: {}", viewPath);
 
-    Validate.isTrue(viewPath.toFile().exists(), "No such file: " + viewId);
+    Validate.isTrue(Files.exists(viewPath), "No such file: " + viewId);
 
-    try (InputStream viewInputStream = Files.newInputStream(viewPath)) {
-      final String translation = EfxTranslator.translateTemplate(new DependencyFactory(sdkRootPath),
-          sdkVersion, viewInputStream);
-      Files.createDirectories(OUTPUT_FOLDER_XSL);
+    final Path filePath =
+        Path.of(NoticeViewerConstants.OUTPUT_FOLDER_XSL.toString(), sdkVersion, viewId + ".xsl");
 
-      final String nameByConvention = viewId + ".xsl";
-      final Path filePath = OUTPUT_FOLDER_XSL.resolve(nameByConvention);
-      try (BufferedWriter writer = new BufferedWriter(new FileWriter(filePath.toFile()))) {
-        writer.write(translation);
+    if (!Files.exists(filePath) || forceBuild) {
+      try (InputStream viewInputStream = Files.newInputStream(viewPath)) {
+        Supplier<String> translator =
+            templateTranslator(sdkRootPath, sdkVersion, viewInputStream, viewId);
+
+        if (forceBuild) {
+          CacheHelper.put(NoticeViewerConstants.NV_CACHE_REGION, translator.get(),
+              new String[] {sdkRootPath.toString(), sdkVersion, viewId});
+        }
+
+        final String translation =
+            CacheHelper.get(translator, NoticeViewerConstants.NV_CACHE_REGION,
+                new String[] {sdkRootPath.toString(), sdkVersion, viewId});
+
+        Files.createDirectories(filePath.getParent());
+        try (BufferedWriter writer = new BufferedWriter(new FileWriter(filePath.toFile()))) {
+          writer.write(translation);
+        }
+
+        logger.debug("Successfully created XSL for view ID [{}] and SDK version [{}]: {}", viewId,
+            sdkVersion, filePath);
       }
-
-      logger.debug("Successfully created XSL for view ID [{}] and SDK version [{}]: {}", viewId,
-          sdkVersion, filePath);
-
-      return filePath;
     }
+
+    return filePath;
   }
 
   /**
